requirements.txt
beautifulsoup4
pytube3
requests